coti revisada 1

+++ modules/cotizacion/controllers/generarCotizacionController.php
@@ -178,58 +178,28 @@ class generarCotizacionController extends Controller{
     public function postEmail(){ 
         $this->postPDF('N');
         Obj::run()->generarCotizacionModel->postTiempoCotizacion();
-        
-        $data = Obj::run()->generarCotizacionModel->getCotizacion();
-        
+                       
         $emailCliente = $data[0]['email'];
         $cliente = $data[0]['nombrecompleto'];
         $emailUser = $data[0]['mail_user'];
         
-        $body ='
-            <h3>Cotización N° '.$data[0]['cotizacion_numero'].'</h3>
-            <h3>Cliente: '.$data[0]['nombrecompleto'].'</h3>
-        <table border="1" style="border-collapse:collapse">
-            <tr>
-                <th style="width:10%">Código</th>
-                <th style="width:40%">Producto</th>
-                <th style="width:10%">Precio</th>
-                <th style="width:10%">Meses</th>
-                <th style="width:10%">Producción</th>
-                <th style="width:10%">Importe</th>
-            </tr>';
-        foreach ($data as $value) {
-            $body .= '<tr>
-                <td style="text-align:center">'.$value['codigo'].'</td>
-                <td>'.$value['producto'].'</td>
-                <td style="text-align:right">'.number_format($value['precio'],2).'</td>
-                <td style="text-align:center">'.$value['cantidad_mes'].'</td>
-                <td style="text-align:right">'.number_format($value['costo_produccion'],2).'</td>
-                <td style="text-align:right">'.number_format($value['importe'],2).'</td>
-            </tr>';
-        }    
-        $body .='</table>';
-        
+        //Html de Cotizacion:
+        $body = $this->getHtmlCotizacion();        
         
         $mail             = new PHPMailer(); // defaults to using php "mail()"
  
 //        $body             = file_get_contents('contents.html');
 //        $body             = eregi_replace("[\]",'',$html);
 
-        $mail->AddReplyTo("name@gmail.com","First Last");
-
-        $mail->SetFrom('name@gmail.com', 'First Last');
-
-        $mail->AddReplyTo("name@gmail.com","First Last");
-
+        $mail->AddReplyTo($emailUser,"VENDEDOR");
+        //$mail->SetFrom('correoempresa@sevend.pe', 'SEVEND');
+        $mail->AddReplyTo("correoempresa@sevend.pe","SEVEND");
         $mail->AddAddress($emailCliente, $cliente);
         $mail->AddAddress($emailUser, $emailUser);
-
-        $mail->Subject    = "PHPMailer Test Subject via mail(), basic";
-
+        $mail->Subject    = "Cotizacion";
         $mail->AltBody    = "To view the message, please use an HTML compatible email viewer!"; // optional, comment out and test
-
+        
         $mail->MsgHTML($body);
-
         $mail->AddAttachment('public/files/cotizacion.pdf');      // attachment
 //        $mail->AddAttachment("public/img/phpmailer_mini.gif"); // attachment
         
@@ -246,11 +216,9 @@ class generarCotizacionController extends Controller{
     
     public function postPDF($n=''){
         $ar = ROOT.'public'.DS.'files'.DS.'cotizacion.pdf';
-        /*se elimina el archivo*/
-        unlink($ar);
-        
-        $data = Obj::run()->generarCotizacionModel->getCotizacion();
-        
+        /*Si existe el archivo, se elimina*/
+        if (file_exists($ar)) unlink($ar);
+               
         $mpdf = new mPDF('c');
 
         $mpdf->mirrorMargins = 1;
@@ -261,39 +229,15 @@ class generarCotizacionController extends Controller{
         $mpdf->defaultfooterfontstyle = B; /* blank, B, I, or BI */
         $mpdf->defaultfooterline = 1; /* 1 to include line below header/above footer */
         
-        $html ='
-        <style>
-            table tr td, table tr th{
-                font-size: 11px;
-            }
-        </style>
-        <h3>Cotización N° '.$data[0]['cotizacion_numero'].'</h3>
-        <h4>Cliente: '.$data[0]['nombrecompleto'].'</h4>
-        <table border="1" style="border-collapse:collapse">
-            <tr>
-                <th style="width:10%">Código</th>
-                <th style="width:20%">Tipo</th>
-                <th style="width:30%">Ubicación</th>
-                <th style="width:30%">Medidas</th>
-                <th style="width:10%">Meses</th>
-                <th style="width:10%">Precio</th>
-                <th style="width:10%">Producción</th>
-                <th style="width:10%">Total</th>
-            </tr>';
-        foreach ($data as $value) {
-            $html .= '<tr>
-                <td style="text-align:center">'.$value['codigo'].'</td>
-                <td>'.$value['elemento'].'</td>
-                <td>'.$value['producto'].'</td>
-                <td>'.$value['dimension_ancho'].' x '.$value['dimension_alto'].' mts</td>
-                <td style="text-align:right">'.$value['cantidad_mes'].'</td>
-                <td style="text-align:center">'.number_format($value['precio'],2).'</td>
-                <td style="text-align:right">'.number_format($value['costo_produccion'],2).'</td>
-                <td style="text-align:right">'.number_format($value['importe'],2).'</td>
-            </tr>';
-        }    
-        $html .='</table>';
-        
+        $mpdf->SetHTMLHeader('<img src="'.ROOT.'public'.DS.'img'.DS.'logotipo.png" width="137" height="68" />','',TRUE);
+        $mpdf->SetHTMLFooter('<table width="100%" style="vertical-align: bottom; font-family: serif; font-size: 8pt; color: #000000; font-weight: bold;"><tr>
+                                <td width="33%"><span style="font-weight: bold;">{DATE j-m-Y}</span></td>
+                                <td width="33%" align="center" style="font-weight: bold;">{PAGENO}/{nbpg}</td>
+                                <td width="33%" style="text-align: right; ">SEVEND.pe</td>
+                             </tr></table>');
+                
+        $html = $this->getHtmlCotizacion();         
+
         $mpdf->WriteHTML($html);
         $mpdf->Output($ar,'F');
         
@@ -308,32 +252,7 @@ class generarCotizacionController extends Controller{
         /*se elimina el archivo*/
         unlink($ar);
         
-        $data = Obj::run()->generarCotizacionModel->getCotizacion();
-        
-        $html ='
-        <h3>Cotización N° '.$data[0]['cotizacion_numero'].'</h3>
-        <h4>Cliente: '.$data[0]['nombrecompleto'].'</h4>
-        <table border="1" style="border-collapse:collapse">
-            <tr>
-                <th style="width:10%">Código</th>
-                <th style="width:40%">Producto</th>
-                <th style="width:10%">Precio</th>
-                <th style="width:10%">Meses</th>
-                <th style="width:10%">Producción</th>
-                <th style="width:10%">Importe</th>
-            </tr>';
-        foreach ($data as $value) {
-            $html .= '<tr>
-                <td style="text-align:center">'.$value['codigo'].'</td>
-                <td>'.$value['producto'].'</td>
-                <td style="text-align:right">'.number_format($value['precio'],2).'</td>
-                <td style="text-align:center">'.$value['cantidad_mes'].'</td>
-                <td style="text-align:right">'.number_format($value['costo_produccion'],2).'</td>
-                <td style="text-align:right">'.number_format($value['importe'],2).'</td>
-            </tr>';
-        }    
-        $html .='</table>';
-        
+        $html = $this->getHtmlCotizacion();
         
         $f=fopen($ar,'wb');
         if(!$f){$data = array('result'=>2);}
@@ -354,6 +273,71 @@ class generarCotizacionController extends Controller{
         $data = Obj::run()->generarCotizacionModel->findCotizacion();
         
         return $data;
+    }
+    
+    public function getHtmlCotizacion(){
+        $data = Obj::run()->generarCotizacionModel->getCotizacion();
+        $html ='
+        <style>           
+           table,h3,h4{font-family:Arial;} 
+           table, table td, table th{font-size:12px;}
+           table{width:100%;}
+           #td2 th{background:#901D78; color:#FFF; height:25px;}
+           #td2 td{font-size:11px;height:25px;}
+        </style>            
+        <table width="100%" border="0" cellpadding="5" cellspacing="3">
+          <tr bgcolor="#901D78">
+            <th colspan="4"><div align="center"><h2 style="color:#FFF;">PRESUPUESTO DE PANELES</h2></div></th>
+          </tr>
+          <tr>
+            <td width="13%"><strong>N° Cotización:</strong></td>
+            <td width="56%"><h3>'.$data[0]['cotizacion_numero'].'</h3></td>
+            <td width="8%"><strong>Fecha:</strong></td>
+            <td width="23%">'.Functions::cambiaf_a_normal($data[0]['fecha_cotizacion']).'</td>
+          </tr>
+          <tr>
+            <td><strong>Cliente:</strong></td>
+            <td colspan="3">'.($data[0]['ruccliente']==''?'':$data[0]['ruccliente'].' - ').$data[0]['razonsocial'].'</td>
+          </tr>
+          <tr>
+            <td><strong>Representante:</strong></td>
+            <td colspan="3">'.($data[0]['numerodocumento']==''?'':$data[0]['numerodocumento'].' - ').$data[0]['nombrecompleto'].'</td>
+          </tr>
+          <tr>
+            <td><strong>Campaña:</strong></td>
+            <td colspan="3">'.$data[0]['nombre_campania'].'</td>          
+          </tr>
+        </table> 
+        <br />
+        <table id="td2" border="1" style="border-collapse:collapse">
+            <tr>
+                <th style="width:9%" >Código</th>
+                <th style="width:10%">Elemento</th>
+                <th style="width:40%">Ubicación</th>
+                <th style="width:5%">Medidas</th>
+                <th style="width:8%">Meses</th>
+                <th style="width:12%">Alquiler</th>
+                <th style="width:12%">Producción</th>
+                <th style="width:12%">Total</th>
+            </tr>';
+        foreach ($data as $value) {
+            $html .= '<tr>
+                <td style="text-align:center">'.$value['codigo'].'</td>
+                <td>'.$value['elemento'].'</td>
+                <td>'.$value['producto'].'</td>
+                <td>'.number_format($value['dimension_ancho']).' x '.number_format($value['dimension_alto']).' mts</td>
+                <td style="text-align:center">'.number_format($value['cantidad_mes']).'</td>
+                <td style="text-align:right">S/.'.number_format($value['precio'],2).'</td>
+                <td style="text-align:right">S/.'.number_format($value['costo_produccion'],2).'</td>
+                <td style="text-align:right">S/.'.number_format($value['importe'],2).'</td>
+            </tr>';
+        }    
+        $html .= '<tr><td colspan="6"></td><td>Importe:</td><td style="text-align:right">S/.'.number_format($data[0]['subtotal'],2).'</td></tr>';
+        $html .= '<tr><td colspan="6"></td><td>Impuesto:</td><td style="text-align:right">S/.'.number_format($data[0]['impuesto'],2).'</td></tr>';
+        $html .= '<tr><td colspan="6"></td><td>Total:</td><td style="text-align:right">S/.'.number_format($data[0]['total'],2).'</td></tr>';
+        
+        $html .='</table>';
+        return $html;
     }
     
 }
